Update the edit curriculum API for super admin and admin user.

+++ curriculum/serializers.py
@@ -140,26 +140,37 @@ class CurriculumDetailSerializer(serializers.ModelSerializer):
 
 class CurriculumDetailUpdateSerializer(serializers.ModelSerializer):
     syllabus = ImageFieldStringAndFile(required=False)
-    subjects = CurriculumSubjectsSerializer(many=True, required=False)
+    # subjects = CurriculumSubjectsSerializer(many=True, required=False)
     class Meta:
         model = Curriculum
-        fields = ['id', 'curriculum_name', 'select_class', 'subjects', 'syllabus', 'discription']
+        fields = ['id', 'curriculum_name', 'select_class', 'syllabus', 'discription']
 
     def update(self, instance, validated_data):
         instance.curriculum_name = validated_data.get('curriculum_name', instance.curriculum_name)
-        instance.select_class = validated_data.get('select_class', instance.select_class)
-        instance.syllabus = validated_data.get('syllabus', instance.syllabus)
-        instance.discription = validated_data.get('discription', instance.discription)
+        instance.select_class = validated_data.get('class_name', instance.select_class)
+        instance.discription = validated_data.get('class_name', instance.discription)
+        instance.syllabus = validated_data.get('class_name', instance.syllabus)
         instance.save()
 
-        subjects_data = validated_data.pop('subjects', [])
+        request = self.context.get('request')
+        if not request:
+            raise ValueError("Request is not in context")
+
+        primary_subjects = []
+        optional_subjects = []
+
+        for key in request.data.keys():
+            if key.startswith('primary_subject'):
+                primary_subjects.append(request.data[key])
+            if key.startswith('optional_subject'):
+                optional_subjects.append(request.data[key])
+
+        if not primary_subjects or not optional_subjects:
+            raise ValueError("Primary or Optional subjects are missing")
 
         Subjects.objects.filter(curriculum_id=instance.id).delete()
 
-        for subject_data in subjects_data:
-            primary_subject = subject_data.get('primary_subject', None)
-            optional_subject = subject_data.get('optional_subject', None)
-
+        for primary_subject, optional_subject in zip(primary_subjects, optional_subjects):
             Subjects.objects.create(
                 curriculum_id=instance,
                 primary_subject=primary_subject.strip() if primary_subject else None,
